Add get_blob_raw for Managed Disk Snapshots
Add ability to read Snapshots taken for Managed Disks.
We add get_blob_raw for the Snapshots which mirrors the code for
Managed Disks themselves via the Storage DiskService.
A common helper file raw_blob_helper.rb with the RawBlobHelper
module is added to be used by the DiskService and SnapshotService.

In addition the EndGetAccess operation must be performed when each read
operation is finished so that an attempt to remove the Snapshot will succeed.
